remove server message from email field
